Fix multiple bugs and improvements v2

+++ app/Http/Controllers/Dashboard/CategoriesController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Dashboard;
+
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Exception;
@@ -17,18 +18,26 @@ class CategoriesController extends Controller
     /**
      * Display a listing of the resource.
      */
+
     public function index()
     {
-        if (!Gate::allows('categories.view')) {
-            abort(403);
-        }
+        // if (!Gate::allows('categories.view')) {
+        //     abort(403);
+        // }
+    if (!auth()->user()->store_id) {
+        return redirect()->route('dashboard')
+            ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+    }
+
         $request = request();
 
         $categories = Category::with('parent')
-        ->withCount(['products as products_number' => function($query){
-           $query->where('status','=','active');
-        }]
-        )->filter(filter: $request->query())->paginate();
+            ->where('store_id', auth()->user()->store_id)
+            ->withCount(
+                ['products as products_number' => function ($query) {
+                    $query->where('status', '=', 'active');
+                }]
+            )->filter(filter: $request->query())->paginate();
         // leftJoin('categories as parents', 'parents.id', '=', 'categories.parent_id')
         //     ->select([
         //         'categories.*',
@@ -49,8 +58,11 @@ class CategoriesController extends Controller
      */
     public function create()
     {
-          Gate::authorize('categories.create');
-
+        //   Gate::authorize('categories.create');
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
         $parents = category::all();
         $category = new Category();
         return view('dashboard.categories.create', compact('parents', 'category'));
@@ -61,7 +73,11 @@ class CategoriesController extends Controller
      */
     public function store(Request $request)
     {
-        Gate::authorize('categories.create');
+        // Gate::authorize('categories.create');
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
         $request->validate(Category::rules());
 
         $request->merge([
@@ -84,12 +100,10 @@ class CategoriesController extends Controller
             $data['image'] = $path;
         } else {
             $data['image'] = null;
-
         }
 
         Category::create($data);
         return to_route('categories.index')->with('success', 'Category Created!');
-
     }
 
     /**
@@ -97,11 +111,14 @@ class CategoriesController extends Controller
      */
     public function show(Category $category)
     {
-              Gate::authorize('categories.view');
+        //   Gate::authorize('categories.view');
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
 
-        
-        return view('dashboard.categories.show',[
-            'category'=>$category
+        return view('dashboard.categories.show', [
+            'category' => $category
         ]);
     }
 
@@ -110,13 +127,15 @@ class CategoriesController extends Controller
      */
     public function edit(string $id)
     {
-                Gate::authorize('categories.update');
-
+        // Gate::authorize('categories.update');
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
         try {
             $category = Category::findOrFail($id);
         } catch (Exception $th) {
             return to_route('categories.index')->with('info', 'Record not found');
-
         }
         // هاتلي كل الداتا من جدول الكاتيجري ماعادا الصف الي 
         //  بتاعه هو ال بينعدل فيه حاليا id  (يعني ميجيبش نفسه )
@@ -137,7 +156,11 @@ class CategoriesController extends Controller
      */
     public function update(Request $request, string $id)
     {
-                        Gate::authorize('categories.update');
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
+        // Gate::authorize('categories.update');
 
         $request->validate(Category::rules($id));
 
@@ -171,25 +194,34 @@ class CategoriesController extends Controller
      */
     public function destroy(string $id)
     {
-                        Gate::authorize('categories.delete');
-
+        // Gate::authorize('categories.delete');
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
         $ssinglecategoryfromDB = Category::findOrFail($id);
         $ssinglecategoryfromDB->delete();
 
         return to_route('categories.index')->with('success', 'Category deleted');
-
     }
     public function trash()
     {
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
         $categories = Category::onlyTrashed()->paginate();
 
 
 
         return view('dashboard.categories.trash', compact('categories'));
-
     }
     public function restore(Request $request, $id)
     {
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
         $category = Category::onlyTrashed()->findOrFail($id);
         $category->restore();
 
@@ -197,20 +229,20 @@ class CategoriesController extends Controller
 
 
         return to_route('categories.trash')->with('success', 'Category restore!');
-
     }
     public function forceDelete($id)
     {
+        if (!auth()->user()->store_id) {
+            return redirect()->route('dashboard')
+                ->with('info', 'Your account is not activated yet. Please contact the administrator to activate your account.');
+        }
         $ssinglecategoryfromDB = Category::onlyTrashed()->findOrFail($id);
         $ssinglecategoryfromDB->forceDelete();
 
         if ($ssinglecategoryfromDB->image) {
             Storage::disk('public')->delete($ssinglecategoryfromDB->image);
-
         }
 
         return to_route('categories.trash')->with('success', 'Category deleted forever !');
-
     }
-
 }
